Refactor SimEng core into a shared library
The core sources now provide a `libsimeng.{so,dylib}` shared library.
There is a `simeng` executable providing the main driver for using
SimEng on the command line, which links to this library. The SimEng
headers which define the API are installed to
`${CMAKE_INSTALL_PREFIX}/include/simeng`.

The unit tests now link to this library as well, so no longer need to
rebuild all of the core SimEng sources.