Fix RAW violation check in LSQ
When a store is committed, the code was previously only checking loads
that have executed (i.e. received all of their data). This meant that
loads which have started prior to the store committing (but not yet
received all of their data) would not be checked.

The fix is to add a hash set to track uncommitted loads that have
started, regardless of whether they have received their data or not,
and check all of them when a store commits.