Add support for instructions that serialize
If renaming an instruction's destination register is not supported,
the instruction is considered to be serializing, and the rename unit
will stall until all prior instructions have retired from the reorder
buffer. Renaming is only supported for a given register type if there
are more physical registers than architectural registers for that
type.