[AArch64] Add support for system registers
System register identifiers are mapped onto a zero-indexed set of
register tags so that they can be treated the same as other registers
when providing operand values and performing writeback. The number of
physical system registers is the same as the number of architectural
system registers, which disables renaming and causes system register
write instructions (msr) to be serializing.