Merge core branch into master (#148)
Within these commits, emulation/Core.cc is updated with a fix for early execution of load instructions that have no addresses and the inorder/Core.cc also received a fix to prevent register reads when the decodeToExecuteBuffer_ is stalled. The former fixed the simulation stalling when SVE load instructions had no addresses to load from due to predication. The latter prevented the incorrect supply of data due to the lacking consideration of multiple execution stall cycles, a decoded instruction would be supplied with an incorrect value whilst the correct value would yet to be forwarded to it from the currently executing instruction.

A fix was also made in RenameUnit.cc that unstalled the input buffer earlier in its functionality, similar to that of other pipeline units, and various break statements were replaced with return statements to better reflect the logic seen in other pipeline units.