Merge linux branch into master (#149)
Within these commits, the read system call and a fix for the readBufferThen() function are implemented. This fix utilises the sequenceId of the SVC instruction that initiated the system call as a way to identify memory read requests made in its readBufferThen() functionality as opposed to its memory access addresses. Before this, incorrect read data was supplied due to overlapping memory access addresses from different instructions.