Merge branch branch into master (#152)
Within these commits, the predict() and update() functions within SimEng's branch predictors have been altered to take a pointer to the instruction as opposed to just the instruction address. This change was made in response to the A64FX branch predictor needing more information, in addition to the instruction's address, namely whether it was a BLR, BL, or RET instruction. Due to the introduction of #include "simeng/Instruction.hh" in BranchPredictor.hh, the removal of #include "simeng/BranchPredictor.hh" in Instruction.hh was required and subsequently the movement of the BranchPrediction struct into Instruction.hh. The A64FX branch predictor, BTB_BWTPredictor, utilises a BTB with global addressing and 4-way association in each of its entries as to reduce contention. A return address stack (RAS) is also used to push addresses onto the stack in the case of a BLR/BL instruction and pop addresses off of the stack in the case of a RET instruction.

The other changes made are to reflect the change in predict() and update() function parameters.