Merge lsq branch into master (#153)
Within these commits, the LSQ is parameterised with the L1 bandwidth and the permitted request/request types per cycle. The LSQ now stores requests made in a secondary requestQueue_ queue where, upon each tick of the unit, a set of requests are made to the chosen data memory interface whilst adhering to the new parameterised values. Currently, SVE instructions are regarded as one request (as stated in the A64FX documentation) due to lacking scatter and gather instructions which are treated as more than one request, unlike their contiguous counterparts. LDP/STP instructions are correctly treated as two requests.

The test files are altered in line with the new functionality of the LSQ.