Merge dispatch branch into master (#155)
Within these commits, a new reservation station structure is used and central/grouped/individual reservation station arrangments can be defined. The ReservationStation struct contains information relating to the number of entries available and a vector of ReservationStationPort structs that define a queue of ready-state instructions and its associated issue port index. A vector of these reservation station structs can be defined by a parameterised port mapping or rsArrangement vector that defines which ports map to which reservation stations (and implicitly how many reservation stations are in use). A dispatch rate is also defined as a parameter of the dispatch unit which defines how many instructions can be dispatched from each reservation station per cycle.

Changes to DispatchIssueUnit.cc reflect the use of these new structs and adherence to dispatchrate_ and a reservation station's capacity attribute.

The RegressionTest.cc file has had the new rsArrangement parameter (TX2 version) of the DispatchIssueUnit.cc added.