Merge execute branch into master (#156)
Within these commits, the notion of operation stalling found in the A64FX processor has been implemented. Unlike pipeline stalling at execution where all instructions flowing into an execution unit are stalled whilst another instruction's operation is executed (e.g. DIV instruction), an operation stall will only stall instructions of similar operations to that of the instruction currently being executed. Such functionality is defined through the use of a parameterised value called a blockingGroup_. This value denotes an instruction group and if an instruction belonging to this group flows into the execute unit, whilst an instruction belonging to that same group is executing, it is appended to an operationsStalled_ queue. Instructions are taken from this queue individually once the previous instruction belonging to the blockingGroup_ group has finished executing.