Re-implementation of Instruction groups (#167)
The content of this PR is centred around the development of new, more user-friendly, instruction groups and their configuration of the SimEng model. The new instruction groups are pre-defined in an ISA specific namespace (InstructionGroups) and latencies can be assigned to these instruction groups within the configuration files. Additionally, the access latency to the L1-cache must now be a fixed value and any variable latencies for load/store operations are implemented within the LSQ.