Add option to print core statistics in YAML format. (#172)
This PR adds a CMake option to print out core stats and build metadata in a YAML format for use in the performance and accuracy script.

Additionally, a segmentation fault caused by speculative instruction fetching was fixed. It occurred when simulating the bude benchmarks on certain host architectures (x86_64). Page sizes might explain why the segmentation fault only occurs on a subset of host architectures.