Merge all ADD with helper and behaviour methods
Helper and behaviour methods for ADD operation was introduced to avoid
repeating similar code in Instruction::execute().
However, addHelper() uses pretty long switch cases as well. So,
improvements on this method should be considered.

Method extendValue() was copied to Instruction_execute.cc from
arch/aarch64/Instruction.cc temporarily to help the process of
simplifying the Instruction::execute().

ADDv() is not finished. This does not work for now.