Improve helper for ADD operation
Changed the control flow of the addHelper() to use metadata of operands
rather than opcode.
This change was done to generalise the helper, so the same function can
be used for all arithmatic instructions.

New enum OperandType was introduced to control type casting when getting
register value.

The header <iostream> was used for debugging purpose. This will be
removed when operandsChecker() is not needed.

ADDv() is removed as it is not used.