Updated circleci pipeline to automatically build documentation (#184)
Amongst several updates to the documentation, this PR alters the CircleCI pipeline to automatically rebuild the documentation pushed to the main branch. It will subsequently push the newly built documentation to the gh-pages branch for deployment to the GitHub page. The CircleCI pipeline's role to build and run SimEng has now been fully transferred to the Jenkins pipeline.