Reverse register rename rewinding order on flush
When SimEng flushes an instruction that sets multiple destinations as
same register, the rewinding of register renaming fails.
This is because the order of applying rewinding by historyTable_ is
wrong.
Ultimately, it keeps wrong physical register in mappingTable_.
Therefore, the order of calling rewind() was reversed to have correct
order of updating mappingTable_ with historyTable_.