Fix data loss bug in FetchUnit::tick()
If the pc_ was not aligned to blockSize boundary and the
fetchBuffer_ was empty, the fetchData would not be copied
but used directly as an optimization. However, if the fetchData
was not enough to start decoding, the function would exit and
the fetchData would be loss.

To fix the bug, the optimization was removed and fetchData is
always copied onto the fetchBuffer_. The optimization did not
provide any performance improvement on the M1 Mac Mini.