Update RegisterValue ctor to take capacity
The ctor that accepts a reference to an array is updated to accept
capacity as a second parameter. This allows us to create RegisterValue
from arrays without the size of RegisterValue being bound to the array
size.

Type traits were introduced in the first ctor to stop it from
accepting pointers as it's first parameter. If pointers were
allowed, it would clash with array ctor in overload resolution.