Use raw pointers instead of shared_ptr in RegisterValue
We are abandoning C++17 memory pool because it is not yet
supported by LLVM/Clang.

I hypothesize that using raw pointers could give a performance
boost as we do not use any atomics. However, this means that
we need to implement our own copy/move constructors.
I have used the copy-and-swap idiom to implement them.