Exception handler refactor (#285)
Within this PR, the exception handler has been refactored to be instantiated once before the simulation begins and stays alive until the simulation ends, rather than be created and destroyed multiple times throughout the simulation.

Additionally, a large portion of an exception handler's logic has been moved to the syscall handler as it is more appropriate to reside there. Such logic is mainly concerned with syscalls independent of the core class. A new communicating route between the syscall handler and the core class has been established to aid this new rework.