Exit and Exit_Group syscalls (#289)
Implemented the exit and exit_group syscalls, along with a mechanism to allow a syscall to update the core's status to idle, once the process state returned from the syscall has been applied.

Also includes a minor tweak to the scheduling logic, meaning that if scheduledProcs_ is empty then a process in waitingProcs_ can jump ahead.