SST compile fix on MacOS (#338)
On more recent versions of MacOS, and more specifically with more recent versions of Apple Clang, SST is unable to be compiled due to unary_function and binary_function types being depricated (since c++11) and now removed from Apple Clang v15. There is a compile flag which re-enabled these types and this has been added to CMakeLists.txt if the user is on MacOS.

These functions are only used in SST, and in more recent versions (i.e. SST v13) they have been replaced - so this issue may only be short term if SimEng's SST support is updated to version 13+