Global siminfo (#343)
This PR looks to replace the use of Config.hh and YAML-CPP with SimInfo.hh and a single header yaml library rapidyaml. The SimInfo class acts as a central point for all model configuration values both defined manually and from a read-in yaml config file.

Improvements to adding new config options have been included through a new expectation struct that places requirements on a specific config option and a default value. Through this, default config files can be created and are especially useful in the test suite.