Implement fabricating accelerators from configs, move accelerators to separate SST components & start implementing serialization
The serialization mechanism does not handle `RegisterValue`s properly -- for now, when using SST, instructions are not serialized. Instead, they are cloned, the clone's pointer is serialized (see `std::unique_ptr::release`), sent, deserialized, and then reinstantiated. This works if the simulation is not run on multiple machines, but should be refactored once serialization is fully implemented.